backend/requirements.txt
Django>=4.2,<5.0
django-jazzmin>=3.0,<4.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0
python-dotenv>=1.0,<2.0
gunicorn>=21.2,<22.0
whitenoise>=6.6,<7.0
